% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{apply_adjustments}
\alias{apply_adjustments}
\title{Apply price adjustments to data}
\usage{
apply_adjustments(data, chart_result, auto_adjust, back_adjust)
}
\description{
Apply price adjustments to data
}
\keyword{internal}
