% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_timestamp}
\alias{as_timestamp}
\title{Convert Date or POSIXct to Unix timestamp}
\usage{
as_timestamp(x, default = NULL)
}
\arguments{
\item{x}{Date object, POSIXct object, or string in YYYY-MM-DD HH:MM:SS format}

\item{default}{Default value to return if x is NULL}
}
\value{
Unix timestamp in seconds
}
\description{
Convert Date or POSIXct to Unix timestamp
}
\keyword{internal}
