% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{build_history_params}
\alias{build_history_params}
\title{Build request parameters for history API}
\usage{
build_history_params(interval, prepost, period, start_date, end_date)
}
\description{
Build request parameters for history API
}
\keyword{internal}
