% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize first letter of string}
\usage{
capitalize(x)
}
\arguments{
\item{x}{String to capitalize}
}
\value{
String with first letter capitalized
}
\description{
Capitalize first letter of string
}
\keyword{internal}
