% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_is_lgl}
\alias{check_is_lgl}
\title{Check if a variable is a logical value}
\usage{
check_is_lgl(x)
}
\arguments{
\item{x}{The variable to check}
}
\value{
The variable \code{x}
}
\description{
Check if a variable is a logical value
}
\keyword{internal}
