% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_names}
\alias{clean_names}
\title{Clean table or vector of names}
\usage{
clean_names(.data, unique = FALSE, minus_to_underscore = FALSE)
}
\arguments{
\item{.data}{Data.frame or vector of column names.}

\item{unique}{Should the variable names be unique?}

\item{minus_to_underscore}{By default \code{-} is replaced with \code{minus}.
This argument replaces the hyphen with \verb{_} (underscore) instead.}
}
\value{
Returns vector if vector or data.frame if data.frame.
}
\description{
Clean table or vector of names
}
