% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{convert_timestamps}
\alias{convert_timestamps}
\title{Convert timestamps to dates in a tibble}
\usage{
convert_timestamps(tibble)
}
\arguments{
\item{tibble}{A tibble potentially containing timestamp columns}
}
\value{
A tibble with timestamps converted to Date objects
}
\description{
Convert timestamps to dates in a tibble
}
\keyword{internal}
