% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{create_historical_df}
\alias{create_historical_df}
\title{Create historical data frame from chart result}
\usage{
create_historical_df(chart_result)
}
\description{
Create historical data frame from chart result
}
\keyword{internal}
