% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{extract_chart_data}
\alias{extract_chart_data}
\title{Extract and validate chart data from response}
\usage{
extract_chart_data(resp_json)
}
\description{
Extract and validate chart data from response
}
\keyword{internal}
