% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{extract_field_value}
\alias{extract_field_value}
\title{Function to extract a field value based on its type}
\usage{
extract_field_value(field_data)
}
\arguments{
\item{field_data}{The field data to extract}
}
\value{
The extracted field value or NULL if the field data is not present
}
\description{
Function to extract a field value based on its type
}
\keyword{internal}
