% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{extract_nested_field}
\alias{extract_nested_field}
\title{Extract a specific field from nested structures}
\usage{
extract_nested_field(col_data, field_name, is_numeric = TRUE)
}
\arguments{
\item{col_data}{List column data}

\item{field_name}{Field name to extract}

\item{is_numeric}{Whether the field is numeric}
}
\value{
Vector of extracted values
}
\description{
Extract a specific field from nested structures
}
\keyword{internal}
