% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_a1_cookie}
\alias{get_a1_cookie}
\title{Get A1 cookie from Yahoo Finance}
\usage{
get_a1_cookie()
}
\value{
A1 cookie string or NULL if not found
}
\description{
Uses curl_chrome110 to get the A1 cookie from Yahoo Finance
}
\keyword{internal}
