% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials.R
\name{get_cashflow}
\alias{get_cashflow}
\title{Get cash flow statement for a ticker}
\usage{
get_cashflow(
  ticker,
  freq = c("annual", "quarterly"),
  start = NULL,
  end = NULL,
  cashflow_keys = NULL,
  pretty = TRUE,
  wide = TRUE,
  proxy = NULL,
  output = c("tibble", "response", "request")
)
}
\arguments{
\item{ticker}{A ticker object created with get_tickers) or a ticker symbol}

\item{freq}{Frequency of data: "annual" or "quarterly" (default "annual")}

\item{start}{Start timestamp (default EOY 2016)}

\item{end}{End timestamp (default current timestamp)}

\item{cashflow_keys}{Vector of cash flow statement keys (default all)}

\item{pretty}{Format column names to be more readable (default TRUE)}

\item{wide}{Return data in wide format (default TRUE)}

\item{proxy}{Optional proxy settings}

\item{output}{Object to return. Can be "tibble", "response", or "request" (default "tibble")}
}
\value{
Either a tibble with cash flow statement data, an httr2 response object, or an httr2 request object
depending on the value of the output argument.
}
\description{
Get cash flow statement for a ticker
}
\examples{
\dontrun{
apple <- get_tickers("AAPL")

# Get annual cash flow statement
cash_flow <- get_cashflow(apple)

# Get quarterly cash flow statement
quarterly_cash_flow <- get_cashflow(apple, freq = "quarterly")
}
}
