% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_crumb}
\alias{get_crumb}
\title{Get crumb using A1 cookie}
\usage{
get_crumb(a1_cookie, proxy = NULL)
}
\arguments{
\item{a1_cookie}{A1 cookie string}

\item{proxy}{Optional proxy settings}
}
\value{
Crumb string or NULL if not found
}
\description{
Get crumb using A1 cookie
}
\keyword{internal}
