% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials.R
\name{get_financials}
\alias{get_financials}
\title{Get all financial statements for a ticker}
\usage{
get_financials(
  ticker,
  freq = "annual",
  start = NULL,
  end = NULL,
  cashflow_keys = NULL,
  balance_keys = NULL,
  income_keys = NULL,
  pretty = TRUE,
  wide = TRUE,
  proxy = NULL,
  output = c("tibble", "response", "request")
)
}
\arguments{
\item{ticker}{A ticker object created with get_tickers) or a ticker symbol}

\item{freq}{Frequency of data: "annual" or "quarterly" (default "annual")}

\item{start}{Start timestamp (default EOY 2016)}

\item{end}{End timestamp (default current timestamp)}

\item{cashflow_keys}{Vector of cash flow statement keys (default all)}

\item{balance_keys}{Balance sheet keys to retrieve (default all)}

\item{income_keys}{Vector of income statement keys (default all)}

\item{pretty}{Format column names to be more readable (default TRUE)}

\item{wide}{Return data in wide format (default TRUE)}

\item{proxy}{Optional proxy settings}

\item{output}{Object to return. Can be "tibble", "response", or "request" (default "tibble")}
}
\value{
A list containing income statement, balance sheet, and cash flow statement,
or if output is "request" or "response", a list of httr2 request or response objects
}
\description{
Get all financial statements for a ticker
}
\examples{
\dontrun{
apple <- get_tickers("AAPL")

# Get all annual financial statements
financials <- get_financial_statements(apple)

# Get all quarterly financial statements
quarterly_financials <- get_financial_statements(apple, freq = "quarterly")
}
}
