% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickers.R
\name{get_tickers_info}
\alias{get_tickers_info}
\title{Get information for multiple tickers}
\usage{
get_tickers_info(
  tickers_obj,
  modules = "summaryProfile",
  output = c("tibble", "response", "request"),
  proxy = NULL
)
}
\arguments{
\item{tickers_obj}{A tickers object created with get_tickers()}

\item{modules}{Modules to retrieve information for (default "summaryProfile")}

\item{output}{Object to return. Can be "tibble", "response", or "request" (default "tibble")}

\item{proxy}{Optional proxy settings}
}
\value{
A list of information for each ticker
}
\description{
Get information for multiple tickers
}
\examples{
\dontrun{
tech_tickers <- get_tickers(c("AAPL", "MSFT", "GOOG"))
tech_info <- get_tickers_info(tech_tickers)
}
}
