% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{has_nested_structure}
\alias{has_nested_structure}
\title{Check if a column contains nested structures with specific fields}
\usage{
has_nested_structure(col_data)
}
\arguments{
\item{col_data}{List column data}
}
\value{
Boolean indicating if the column contains nested structures
}
\description{
Check if a column contains nested structures with specific fields
}
\keyword{internal}
