% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{null-coalesce-operator}
\alias{null-coalesce-operator}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Fallback value if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Null coalescing operator
}
\keyword{internal}
