% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_asset_profile}
\alias{parse_asset_profile}
\title{Parse asset profile module data}
\usage{
parse_asset_profile(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse asset profile module data
}
\keyword{internal}
