% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_calendar_events}
\alias{parse_calendar_events}
\title{Parse calendar events module data}
\usage{
parse_calendar_events(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse calendar events module data
}
\keyword{internal}
