% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_cashflow_statement}
\alias{parse_cashflow_statement}
\title{Parse cash flow statement module data}
\usage{
parse_cashflow_statement(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse cash flow statement module data
}
\keyword{internal}
