% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_default_key_statistics}
\alias{parse_default_key_statistics}
\title{Parse default key statistics module data}
\usage{
parse_default_key_statistics(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse default key statistics module data
}
\keyword{internal}
