% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_earnings}
\alias{parse_earnings}
\title{Parse earnings module data}
\usage{
parse_earnings(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A list of tibbles with parsed data
}
\description{
Parse earnings module data
}
\keyword{internal}
