% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_earnings_history}
\alias{parse_earnings_history}
\title{Parse earnings history module data}
\usage{
parse_earnings_history(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse earnings history module data
}
\keyword{internal}
