% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_earnings_trend}
\alias{parse_earnings_trend}
\title{Parse earnings trend module data}
\usage{
parse_earnings_trend(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse earnings trend module data
}
\keyword{internal}
