% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_esg_scores}
\alias{parse_esg_scores}
\title{Parse ESG scores module data}
\usage{
parse_esg_scores(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse ESG scores module data
}
\keyword{internal}
