% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_fund_ownership}
\alias{parse_fund_ownership}
\title{Parse fund ownership module data}
\usage{
parse_fund_ownership(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse fund ownership module data
}
\keyword{internal}
