% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_fund_profile}
\alias{parse_fund_profile}
\title{Parse fund profile module data}
\usage{
parse_fund_profile(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse fund profile module data
}
\keyword{internal}
