% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_futures_chain}
\alias{parse_futures_chain}
\title{Parse futures chain module data}
\usage{
parse_futures_chain(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse futures chain module data
}
\keyword{internal}
