% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_generic_module}
\alias{parse_generic_module}
\title{Parse generic module data that requires simple unnesting}
\usage{
parse_generic_module(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse generic module data that requires simple unnesting
}
\keyword{internal}
