% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_income_statement}
\alias{parse_income_statement}
\title{Parse income statement module data}
\usage{
parse_income_statement(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse income statement module data
}
\keyword{internal}
