% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_index_trend}
\alias{parse_index_trend}
\title{Parse index trend module data}
\usage{
parse_index_trend(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse index trend module data
}
\keyword{internal}
