% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_insider_holders}
\alias{parse_insider_holders}
\title{Parse insider holders module data}
\usage{
parse_insider_holders(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse insider holders module data
}
\keyword{internal}
