% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_insider_transactions}
\alias{parse_insider_transactions}
\title{Parse insider transactions module data}
\usage{
parse_insider_transactions(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse insider transactions module data
}
\keyword{internal}
