% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_institution_ownership}
\alias{parse_institution_ownership}
\title{Parse institution ownership module data}
\usage{
parse_institution_ownership(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse institution ownership module data
}
\keyword{internal}
