% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_module_data}
\alias{parse_module_data}
\title{Info.R - Module for retrieving and parsing Yahoo Finance quoteSummary data}
\usage{
parse_module_data(result_data, module_name)
}
\arguments{
\item{result_data}{The result data from Yahoo Finance API}

\item{module_name}{The name of the module to parse}
}
\value{
A tibble with parsed data
}
\description{
This file contains functions for parsing Yahoo Finance quoteSummary API data
into tidy tibble formats using tidyverse functions to handle nested data.
Parse a module's data into a tidy tibble format
}
\keyword{internal}
