% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_ownership_data}
\alias{parse_ownership_data}
\title{Parse ownership data module}
\usage{
parse_ownership_data(result_data, list_field = "ownershipList")
}
\arguments{
\item{result_data}{The data to parse}

\item{list_field}{The field containing the list of ownership data}
}
\value{
A tibble with parsed data
}
\description{
Parse ownership data module
}
\keyword{internal}
