% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_recommendation_trend}
\alias{parse_recommendation_trend}
\title{Parse recommendation trend module data}
\usage{
parse_recommendation_trend(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse recommendation trend module data
}
\keyword{internal}
