% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_sec_filings}
\alias{parse_sec_filings}
\title{Parse SEC filings module data}
\usage{
parse_sec_filings(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse SEC filings module data
}
\keyword{internal}
