% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{parse_ticker_symbols}
\alias{parse_ticker_symbols}
\title{Parse Ticker Symbols}
\usage{
parse_ticker_symbols(symbols)
}
\arguments{
\item{symbols}{A character vector of ticker symbols}
}
\value{
A character vector of parsed ticker symbols
}
\description{
Parse Ticker Symbols
}
\keyword{internal}
