% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_up_down_history}
\alias{parse_up_down_history}
\title{Parse upgrade downgrade history module data}
\usage{
parse_up_down_history(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse upgrade downgrade history module data
}
\keyword{internal}
