% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickers.R
\name{print.yf_ticker}
\alias{print.yf_ticker}
\title{Format ticker object print output}
\usage{
\method{print}{yf_ticker}(x, ...)
}
\arguments{
\item{x}{The ticker object}

\item{...}{Additional arguments passed to print}
}
\value{
The ticker object (invisibly)
}
\description{
Format ticker object print output
}
