% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{process_calendar_earnings}
\alias{process_calendar_earnings}
\title{Process earnings data in calendar events}
\usage{
process_calendar_earnings(tbl, earnings_data)
}
\arguments{
\item{tbl}{The tibble to update}

\item{earnings_data}{The earnings data to process}
}
\value{
Updated tibble with processed earnings data
}
\description{
Process earnings data in calendar events
}
\keyword{internal}
