% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{process_current_quarter}
\alias{process_current_quarter}
\title{Process current quarter estimate data}
\usage{
process_current_quarter(earnings_chart, result_tibbles = list())
}
\arguments{
\item{earnings_chart}{The earnings chart data}

\item{result_tibbles}{The list of result tibbles to update}
}
\value{
Updated result_tibbles list with processed data
}
\description{
Process current quarter estimate data
}
\keyword{internal}
