% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{process_earnings_metadata}
\alias{process_earnings_metadata}
\title{Process metadata from earnings data}
\usage{
process_earnings_metadata(base_tbl, result_tibbles = list())
}
\arguments{
\item{base_tbl}{The base tibble containing metadata}

\item{result_tibbles}{The list of result tibbles to update}
}
\value{
Updated result_tibbles list with metadata
}
\description{
Process metadata from earnings data
}
\keyword{internal}
