% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{process_nested_column}
\alias{process_nested_column}
\title{Process a single nested column}
\usage{
process_nested_column(tibble, col, col_prefix)
}
\arguments{
\item{tibble}{The tibble containing the column}

\item{col}{The column name}

\item{col_prefix}{Prefix to use for new columns}
}
\value{
Updated tibble with processed column
}
\description{
Process a single nested column
}
\keyword{internal}
