% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials.R
\name{process_timeseries_data}
\alias{process_timeseries_data}
\title{Process timeseries data from the fundamentals-timeseries endpoint}
\usage{
process_timeseries_data(result_data, pretty = TRUE, wide = TRUE)
}
\arguments{
\item{result_data}{List of timeseries results}

\item{pretty}{Format column names to be more readable (default TRUE)}

\item{wide}{Return data in wide format (default TRUE)}
}
\value{
A tidy tibble with processed timeseries data
}
\description{
Process timeseries data from the fundamentals-timeseries endpoint
}
\keyword{internal}
