% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_timestamp_arg}
\alias{process_timestamp_arg}
\title{Process Timestamp Argument
Converts a character string, timestamp, or date to a unix timestamp}
\usage{
process_timestamp_arg(timestamp = NULL, default = NULL)
}
\arguments{
\item{timestamp}{A character string that can be converted to a unix timestamp, timestamp, or date}

\item{default}{A default value to return if timestamp is NULL}
}
\value{
A unix timestamp. Default is current time.
}
\description{
Process Timestamp Argument
Converts a character string, timestamp, or date to a unix timestamp
}
\keyword{internal}
