% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{read_auth_file}
\alias{read_auth_file}
\title{Read Auth File}
\usage{
read_auth_file(path = NULL, refresh = FALSE)
}
\arguments{
\item{path}{Path to the authentication file}

\item{refresh}{Whether to refresh the auth file}
}
\description{
Read Auth File
}
\keyword{internal}
