% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{req_add_auth}
\alias{req_add_auth}
\title{Get Yahoo Finance authentication (crumb)}
\usage{
req_add_auth(req, proxy = NULL, refresh = FALSE, path = NULL)
}
\arguments{
\item{req}{Request object}

\item{proxy}{Optional proxy settings}

\item{refresh}{Logical. If TRUE, force a refresh of the crumb.}
}
\value{
Request object with added authentication
}
\description{
Tries multiple methods to get a crumb:
\enumerate{
\item Check YAHOO_FINANCE_CRUMB environment variable
\item Get A1 cookie and use it to fetch a crumb
}
}
\keyword{internal}
