% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr2.R
\name{req_add_headers}
\alias{req_add_headers}
\title{Add headers to a request}
\usage{
req_add_headers(req)
}
\arguments{
\item{req}{Request object}
}
\value{
Request object with added headers
}
\description{
Add headers to a request
}
\keyword{internal}
