% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unix_to_date}
\alias{unix_to_date}
\title{Convert Unix timestamp to date}
\usage{
unix_to_date(timestamp)
}
\arguments{
\item{timestamp}{Unix timestamp in seconds (can be a vector)}
}
\value{
A Date object or vector of Date objects
}
\description{
Convert Unix timestamp to date
}
\keyword{internal}
