% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unix_to_datetime}
\alias{unix_to_datetime}
\title{Convert Unix timestamp to datetime}
\usage{
unix_to_datetime(timestamp)
}
\arguments{
\item{timestamp}{Unix timestamp in seconds (can be a vector)}
}
\value{
A POSIXct object or vector of POSIXct objects
}
\description{
Convert Unix timestamp to datetime
}
\keyword{internal}
