% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{user_agents}
\alias{user_agents}
\title{Static user agents for yfinance package}
\format{
An object of class \code{character} of length 8.
}
\usage{
user_agents
}
\description{
Static user agents for yfinance package
}
\keyword{internal}
