% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_cashflow_keys}
\alias{valid_cashflow_keys}
\title{Data file containing possible cash flow statement keys from Yahoo Finance API}
\format{
A character vector containing 131 possible cash flow statement keys
}
\source{
Yahoo Finance API
}
\usage{
valid_cashflow_keys
}
\description{
A dataset containing the possible keys that can be found in cash flow statement data
retrieved from Yahoo Finance's API. These keys are used to identify and extract
specific cash flow line items.
}
\examples{
# View all possible cash flow keys
data(valid_cashflow_keys)
head(valid_cashflow_keys)

# Check if a specific key exists
"OperatingCashFlow" \%in\% valid_cashflow_keys

}
\keyword{datasets}
