% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_income_keys}
\alias{valid_income_keys}
\title{Data file containing possible income statement keys from Yahoo Finance API}
\format{
A character vector containing 103 possible income statement keys
}
\source{
Yahoo Finance API
}
\usage{
valid_income_keys
}
\description{
A dataset containing the possible keys that can be found in income statement data
retrieved from Yahoo Finance's API. These keys are used to identify and extract
specific income statement line items.
}
\examples{
# View all possible income statement keys
data(valid_income_keys)
head(valid_income_keys)

# Check if a specific key exists
"NetIncome" \%in\% valid_income_keys

}
\keyword{datasets}
