% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_frequency}
\alias{validate_frequency}
\title{Validate frequency}
\usage{
validate_frequency(frequency)
}
\arguments{
\item{frequency}{A character string specifying the frequency (annual, quarterly)}
}
\value{
The validated frequency
}
\description{
Validate frequency
}
\keyword{internal}
