% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_interval}
\alias{validate_interval}
\title{Validate date interval}
\usage{
validate_interval(interval)
}
\arguments{
\item{interval}{A character string specifying the interval (1m, 2m, 5m, 15m, 30m, 60m, 90m, 1h, 1d, 5d, 1wk, 1mo, 3mo)}
}
\value{
The validated interval
}
\description{
Validate date interval
}
\keyword{internal}
