% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_period}
\alias{validate_period}
\title{Validate date period}
\usage{
validate_period(period)
}
\arguments{
\item{period}{A character string specifying the period (1d, 5d, 1mo, 3mo, 6mo, 1y, 2y, 5y, 10y, ytd, max)}
}
\value{
The validated period
}
\description{
Validate date period
}
\keyword{internal}
