% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{yf_base_url}
\alias{yf_base_url}
\title{Constants for Yahoo Finance API}
\format{
An object of class \code{character} of length 1.
}
\usage{
yf_base_url
}
\description{
Constants for Yahoo Finance API
}
\keyword{internal}
