% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials.R
\name{get_income_statement}
\alias{get_income_statement}
\title{Get income statement for a ticker}
\usage{
get_income_statement(
  ticker,
  freq = c("annual", "quarterly"),
  start = NULL,
  end = NULL,
  income_keys = NULL,
  pretty = TRUE,
  wide = TRUE,
  proxy = NULL,
  output = c("tibble", "response", "request")
)
}
\arguments{
\item{ticker}{A ticker object created with get_tickers) or a ticker symbol}

\item{freq}{Frequency of data: "annual" or "quarterly" (default "annual")}

\item{start}{Start timestamp (default EOY 2016)}

\item{end}{End timestamp (default current timestamp)}

\item{income_keys}{Vector of income statement keys (default all)}

\item{pretty}{Format column names to be more readable (default TRUE)}

\item{wide}{Return data in wide format (default TRUE)}

\item{proxy}{Optional proxy settings}

\item{output}{Object to return. Can be "tibble", "response", or "request" (default "tibble")}
}
\value{
Either a tibble with income statement data, an httr2 response object, or an httr2 request object
depending on the value of the output argument.
}
\description{
Get income statement for a ticker
}
\examples{
\dontrun{
apple <- get_tickers("AAPL")

# Get annual income statement
income_stmt <- get_income_statement(apple)

# Get quarterly income statement
quarterly_income <- get_income_statement(apple, freq = "quarterly")
}
}
