% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_info}
\alias{get_info}
\title{Get basic information about a ticker}
\usage{
get_info(
  ticker,
  modules = "summaryProfile",
  output = c("tibble", "response", "request"),
  proxy = NULL
)
}
\arguments{
\item{ticker}{A ticker name or ticker object created with get_tickers).}

\item{modules}{A character vector of modules to request from the API}

\item{output}{The output format for the request (tibble, response, request)}

\item{proxy}{A character string specifying the proxy URL}
}
\value{
A list with company information
}
\description{
Get basic information about a ticker
}
\examples{
\dontrun{
apple <- get_tickers("AAPL")
apple_info <- get_info(apple)
}
}
