% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickers.R
\name{get_tickers_financials}
\alias{get_tickers_financials}
\title{Get all financial statements for multiple tickers}
\usage{
get_tickers_financials(
  tickers_obj,
  freq = c("annual", "quarterly"),
  start_timestamp = NULL,
  end_timestamp = NULL,
  cash_flow_keys = NULL,
  balance_keys = NULL,
  income_keys = NULL,
  pretty = TRUE,
  wide = TRUE,
  proxy = NULL,
  output = c("tibble", "response", "request")
)
}
\arguments{
\item{tickers_obj}{A tickers object created with get_tickers()}

\item{freq}{Frequency of data: "annual" or "quarterly" (default "annual")}

\item{start_timestamp}{Start timestamp for data}

\item{end_timestamp}{End timestamp for data}

\item{cash_flow_keys}{Cash flow statement keys (default all)}

\item{balance_keys}{Balance sheet keys (default all)}

\item{income_keys}{Income statement keys (default all)}

\item{pretty}{Format column names to be more readable (default TRUE)}

\item{wide}{Return data in wide format (default TRUE)}

\item{proxy}{Optional proxy settings}

\item{output}{Object to return. Can be "tibble", "response", or "request" (default "tibble")}
}
\value{
A nested list containing financial statements for each ticker
}
\description{
Get all financial statements for multiple tickers
}
\examples{
\dontrun{
tech_tickers <- get_tickers(c("AAPL", "MSFT", "GOOG"))
tech_financials <- get_tickers_financials(tech_tickers)
}
}
