% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_balance_sheet}
\alias{parse_balance_sheet}
\title{Parse balance sheet module data}
\usage{
parse_balance_sheet(result_data)
}
\arguments{
\item{result_data}{The data to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse balance sheet module data
}
\keyword{internal}
