% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{parse_module_data}
\alias{parse_module_data}
\title{Parse a module's data into a tidy tibble format}
\usage{
parse_module_data(result_data, module_name)
}
\arguments{
\item{result_data}{The result data from Yahoo Finance API}

\item{module_name}{The name of the module to parse}
}
\value{
A tibble with parsed data
}
\description{
Parse a module's data into a tidy tibble format
}
\keyword{internal}
