% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{should_message}
\alias{should_message}
\title{Control message frequency}
\usage{
should_message(msg, interval = 8, timestamp_file = NULL)
}
\arguments{
\item{msg}{The message to display}

\item{interval}{Time interval in hours between showing messages}

\item{timestamp_file}{Name used to ID the timestamp file (defaults to a tempfile)}
}
\value{
NULL invisibly
}
\description{
Shows a message only once every \code{interval} hours
}
\keyword{internal}
